feat(Asset System): Added functionality of serializing and deserializing asset metadata registries
When assets gets imported for the first time, the engine imports it and the asset's metadata are serialized, when the asset
is imported again, it looks for the asset handle in the serialized file and uses that instead. Therefore an asset once imported
will have the same handle throughout its life.